requirements.txt
discord.py
pytest_mock
pytest[asyncio]
pytest-asyncio
python-dotenv
